Moved compressor to utils and made its methods public
